requirements_no_conda.txt
#bokeh==2.3.0
bokeh
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
cytoolz==0.11.0
#dask==2021.3.0
#distributed==2021.3.0
distributed
fastavro==1.3.2
#fastrlock==0.5
fastrlock
Flask==1.1.2
fsspec==0.8.7
HeapDict==1.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
#llvmlite==0.34.0
llvmlite
#locket==0.2.0
locket
MarkupSafe==1.1.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.2.*
mkl-service==2.4.*
#msgpack==1.0.0
msgpack
#numba==0.51.2
#numba>=0.54
numba
#numpy==1.19.2
numpy
#nvtx==0.2.3
nvtx
#olefile==0.46
olefile
packaging==20.9
pandas==1.1.4
partd==1.1.0
Pillow==8.1.2
pip==21.0.1
#protobuf==3.13.0
protobuf
psutil==5.8.0
#pyarrow==1.0.1
pyarrow
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
#pytz==2021.1
pytz
PyYAML==5.4.1
requests==2.25.1
scipy==1.5.3
#setuptools==52.0.0.post20210125
setuptools
six==1.15.0
#sortedcontainers==2.3.0
sortedcontainers
tblib==1.6.0
toolz==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.1
#treelite==1.0.0
treelite
#treelite-runtime==1.0.0
treelite-runtime
#typing-extensions==3.7.4.3
typing-extensions
urllib3==1.26.3
Werkzeug==1.0.1
wheel==0.36.2
#zict==2.0.0
zict
